requirements.txt
# Core
python-telegram-bot==21.6
websockets==12.0
aiohttp==3.10.5
python-dotenv==1.0.1

# ML — optional but strongly recommended
# bot runs without these until you have enough data anyway
scikit-learn==1.5.2
numpy==1.26.4
joblib==1.4.2
